Move cfp tasks to own file
